% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxout.R
\name{boxout}
\alias{boxout}
\title{Outlier detection and batch-effect investigation by boxplots}
\usage{
boxout(x)
}
\arguments{
\item{x}{A matrix of array intensities, samples by row.}
}
\value{
An object of class "boxout"
}
\description{
Compared the empirical distribution function of each array's intensities
to the (smoothed) ecdf of all arrays' intensities pooled. The comparison
is done by the Kolmogorov--Smirnov statistic where we ignore the presence of
ties.
}
