% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{corrected}
\alias{corrected}
\title{Background correction of expression values}
\usage{
corrected(data, negative_controls)
}
\arguments{
\item{data}{matrix or LumiBatch, observations by columns, probes by rows}

\item{negative_controls}{matrix of negative probe intensities for background
correction.}
}
\description{
Normal-exponential background correction based on negative probes
}
