% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens.R
\name{dens}
\alias{dens}
\title{Estimate array-wise densities}
\usage{
dens(x)
}
\arguments{
\item{x}{A matrix of array intensities, samples by row.}
}
\description{
This metod basically just runs \code{density()} per row in the input
matrix and packages this in an object of class \code{dens}
}
