% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filtered}
\alias{filtered}
\title{Filtering of bad quality/non-detectable probes}
\usage{
filtered(data, pval = 0.01, fval = 0.01, verbose = FALSE)
}
\arguments{
\item{data}{a LumiBatch}

\item{pval}{a threshold for non-detection of a signal. a signal with detection
p-value above \code{pval} will be discarded.}

\item{fval}{the smallest proportion of observations in which a signal should be
detected (as defined by \code{pval}). If \code{fval} is 0.5, there must be a
detectable signal in half the observations or the probe will be discarded.}

\item{verbose}{print info about filtered-out probes? T/F}
}
\description{
Filters out probes that have a signal below a specified detection threshold
and that arent expressed in at least some specified fraction of observations.
Also removes probes of "poor" or "no match" quality. See reference
}
\references{
Günther C, Holden M, Holden L. Preprocessing of gene-expression data related to breast cancer diagnosis: SAMBA/35/14. \url{https://www.nr.no/files/samba/smbi/note2015SAMBA3514preprocessing.pdf}
}
