% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{gene_names}
\alias{gene_names}
\title{Get gene name mapping}
\usage{
gene_names(data)
}
\arguments{
\item{data}{gene expression matrix or LumiBatch with nuIDs as
its \code{rowname()}s.}
}
\description{
Maps gene symbols to probe nuIDs for your data
}
