% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\title{Highlight an observation in four different views}
\usage{
highlight(observation, pca, box, dens, ma)
}
\arguments{
\item{observation}{the name of the observation to highligh}

\item{pca}{an object of class "prcout" where the observation occors}

\item{box}{an object of class "boxout" where the observation occors}

\item{dens}{an object of class "dens" where the observation occors}

\item{ma}{an object of class "mapout" where the observation occors}
}
\description{
Provides linking of plots for one observation in the four different
models we provide. Puts the plots in a grid an calls \code{plot(x, highlight=observation)}
for each of the four models.
}
