% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_hist.R
\name{hla_hist}
\alias{hla_hist}
\alias{blood_probes}
\title{Blood type probes}
\source{
The Genomics Core Facility, NTNU \url{https://www.ntnu.edu/dmf/gcf}
}
\usage{
hla_hist()

blood_probes()
}
\value{
A character vector of nuIDs. If you for some reason want to
find the Illumina IDs behind these nuIDs you can use \code{names()} on the
returned vector.
}
\description{
Returns a list of probes that cover HLA genes. These are usually strongly
expressed and it may sometimes be desirable to remove them before
analysis. The function \code{blood_probes()} is a wrapper to this
function with a friendlier name.
}
\section{Functions}{
\itemize{
\item \code{blood_probes}: A more descriptive name for hla_hist
}}

\examples{
## nuIDs:
hla_hist()

## Illumina IDs:
names(hla_hist())
}
