% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_measures.R
\docType{data}
\name{lab_thresholds}
\alias{lab_thresholds}
\title{Grounds for excluding a sample based on lab measures}
\format{An object of class \code{character} of length 4.}
\usage{
lab_thresholds
}
\description{
We have several quality control measurements from the lab; when we are
uncertain whether an array looks bad we can inspect these measures and
get a sense of whether something has gone wrong. We use these cutoffs to
say that an array has gone wrong:
\itemize{
 \item{\strong{RIN < 7:}}{ An array with RIN less than 7 is considered bad.}
 \item{\strong{260/280 RNA ratio < 2:}}{ An array with 260/280 less than 2 is considered bad.}
 \item{\strong{260/230 RNA ratio < 1.7:}}{ An array with 260/230 less than 1.7 is considered bad.}
 \item{\strong{Ng/ul RNA outside of (50, 500):}}{ Well-behaved arrays should have
 50 < Ng/ul RNA < 500.}
}
\strong{NB} that perfectly well-behaved arrays might have lab measures outside
of these limits, we never exclude arrays based only on lab measures.
}
\keyword{datasets}
