% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_measures.R
\docType{data}
\name{lab_variables}
\alias{lab_variables}
\title{Vector of the typical column names for lab variables}
\format{An object of class \code{character} of length 7.}
\usage{
lab_variables
}
\description{
Vector of the typical column names for lab variables
}
\keyword{datasets}
