% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapout.R
\name{mapout}
\alias{mapout}
\title{MA-plot-based outlier detection}
\usage{
mapout(x)
}
\arguments{
\item{x}{a matrix of \strong{log transformed, raw} gene expression values
(not fold change). Observations by row, probes by columns.}
}
\value{
an object of class \code{mapout}
}
\description{
Computes M and A statistics between each array and a probe-wise median array. If an
array is well-behaved, there should be no trend in M as a function of A. This
trendiness is estimated by mutual information.
}
