% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalized}
\alias{normalized}
\title{Normalize gene expression values}
\usage{
normalized(data)
}
\arguments{
\item{data}{matrix or LumiBatch, observations by columns, probes by rows}
}
\description{
Does quantile normalization of gene epression values
}
