% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens.R
\name{plot.dens}
\alias{plot.dens}
\title{Plot method for \code{dens} objects}
\usage{
\method{plot}{dens}(obj, batch = NULL, highlight = NULL)
}
\arguments{
\item{obj}{an object of class "dens"}

\item{batch}{a vector indicating which batch a sample belongs to. used for coloring by batch.}

\item{highlight}{sample names of samples to highligth in red on the plot. useful for inspecting specific outliers}
}
\description{
Plots all the densities in the \code{dens}-object in a single plot.
}
