% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapout.R
\name{plot.mapout}
\alias{plot.mapout}
\title{Plot method for \code{mapout} objects}
\usage{
\method{plot}{mapout}(x, nout = 1, highlight = NULL, lineup = F,
  subsample = T, ncol = ifelse(is.null(highlight), nout, length(highlight)),
  ...)
}
\arguments{
\item{x}{object of class \code{mapout}}

\item{nout}{now many outliers to show MAplots for}

\item{highlight}{optional character vector with names of samles to show in the plot.
Overrides the standard plotting of outliers}

\item{lineup}{show \code{nout} random samples for comparison? defaults to false.}

\item{subsample}{show random subset of 2500 probes for speed. Defaults to true}

\item{ncol}{number of columns in output plot}

\item{...}{unused}
}
\description{
Plots the MA-plots \code{nout} "worst outliers" in a grid of \code{ncol}
columns and plots a loess-smoothed line over the data. You want this line
to be as horizontal as possible, but a certain banana-shape is not uncommon.
The loess-smoothing might be slow, and you might end up with a lot of points
in your plot. To work around this, use \code{subsample=T} (default) to work
examine a random subset of 2500 probes.
}
