% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcout.R
\name{plot.prcout}
\alias{plot.prcout}
\title{Plot method for \code{prcout} objects}
\usage{
\method{plot}{prcout}(obj, ...)
}
\arguments{
\item{obj}{object of class \code{prcout}}

\item{...}{passed to predict(obj, ...) to label outliers}
}
\description{
Plots the two first PCs of the data with contour lines corresponding
to the mahalanobis distance to the data center (as defined by the inner mass
of the data)
}
\seealso{
\code{\link{prcout}}, \code{\link{predict.prcout}}
}

