% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prcout.R
\name{prcout}
\alias{prcout}
\title{PCA-based outlier detection}
\usage{
prcout(x, sdev = 2)
}
\arguments{
\item{x}{a numerical matrix with samples by row, measurements by column}

\item{sdev}{Number of standard deviations (in mahalanobis distance) that an
 observation should be from the center to be called an outlier}
}
\value{
an object of class \code{prcout}. Can be plottet with plot(obj)
}
\description{
Takes a matrix of samples x measurements and looks for outliers in the two
first principal components of the data as defined by mahalanobis distance
to the center of the data in number of standard deviations
}

