% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxout.R
\name{predict.boxout}
\alias{predict.boxout}
\title{Predict method for \code{boxout} objects}
\usage{
\method{predict}{boxout}(obj, sdev = 3)
}
\arguments{
\item{sdev}{Number of standard deviations (in KS statistic) that an
observation should be larger than the mean KS statistic in order to be
considered an outlier}
}
\value{
a logical vector indicating outlier Y/N as defined by the \code{sdev}
 parameter.
}
\description{
Provides a prediction for whether an observation is a suspected outlier
}
