% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapout.R
\name{predict.mapout}
\alias{predict.mapout}
\title{Predict method for \code{mapout} objects}
\usage{
\method{predict}{mapout}(object, sdev = 3, ...)
}
\arguments{
\item{object}{object of class "mapout"}

\item{sdev}{Number of standard deviations (in mutual information) that an
observation should be from the center to be called an outlier}

\item{...}{unused}
}
\value{
a character vector of row names for the outliers as defined by the \code{sdev}
 parameter.
}
\description{
Provides a prediction for whether an observation is a suspected outlier
}
