% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcout.R
\name{predict.prcout}
\alias{predict.prcout}
\title{Predict method for \code{prcout} objects}
\usage{
\method{predict}{prcout}(obj, sdev = 3)
}
\arguments{
\item{obj}{object of class "prcout"}

\item{sdev}{Number of standard deviations (in mahalanobis distance) that an
observation should be from the center to be called an outlier}
}
\value{
a character vector of row names for the outliers as defined by the \code{sdev}
 parameter.
}
\description{
Provides a prediction for whether an observation is a suspected outlier
}
