% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{probe_aggregated}
\alias{probe_aggregated}
\title{Aggregates by annotation-based filtering}
\usage{
probe_aggregated(data, verbose = FALSE)
}
\arguments{
\item{data}{a LumiBatch}

\item{verbose}{print info about filtered-out probes? T/F}
}
\description{
Aggregates gene expression values across probes by the "lumiHumanAll" annotation
package
}
