% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcout.R
\name{plot.prcout}
\alias{plot.prcout}
\title{Plot method for \code{prcout} objects}
\usage{
\method{plot}{prcout}(obj, highlight = NULL, ...)
}
\arguments{
\item{obj}{object of class \code{prcout}}

\item{highlight}{optional character vector with names of samles to highlight in the plot.
Overrides the highlighting of outliers}

\item{...}{passed to predict(obj, ...) to label outliers}
}
\description{
Plots the two first PCs of the data with contour lines corresponding
to the mahalanobis distance to the data center (as defined by the inner mass
of the data). Highlights outliers or the samples defined in \code{highlight=}
}
\seealso{
\code{\link{prcout}}, \code{\link{predict.prcout}}
}

