% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcout.R
\name{prcout}
\alias{prcout}
\title{PCA-based outlier detection}
\usage{
prcout(x, prob = 0.01)
}
\arguments{
\item{x}{a numerical matrix with samples by row, measurements by column}

\item{prob}{How unlikely should a data point at least be in order to not be
considered part of the "center mass" of the data. Translated to k in
Chebyshev's inequality P(|Z| >= k) =< 1/k^2 and applied to the two
first PCs.}
}
\value{
an object of class \code{prcout}. Can be plotted with plot(obj)
}
\description{
Takes a matrix of samples x measurements and looks for outliers in the two
first principal components of the data as defined by mahalanobis distance
to the center of the data in number of standard deviations
}
