% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxout.R
\name{predict.boxout}
\alias{predict.boxout}
\title{Predict method for \code{boxout} objects}
\usage{
\method{predict}{boxout}(object, sdev = 3, ...)
}
\arguments{
\item{object}{an object of class "boxout"}

\item{sdev}{Number of standard deviations (in KS statistic) that an
observation should be larger than the mean KS statistic in order to be
considered an outlier}

\item{...}{unused}
}
\value{
a vector of names for the outlying observations as defined by the \code{sdev}
 parameter.
}
\description{
Provides a prediction for whether an observation is a suspected outlier
}
